ASKMYPDF/requirement.txt
streamlit==1.44.1
langchain-community==0.3.22
langchain-google-genai==2.1.3
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.8
openai==1.72.0
python-dotenv==1.1.0
